feature(api): env 적용 및 상세페이지 api 추가
