+++ package.json
@@ -44,7 +44,7 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "^24.0.12",
+    "@types/node": "^24.0.14",
     "tsconfig-paths-webpack-plugin": "^4.2.0",
     "typescript": "^4.9.5"
   }
