+++ src/custom.d.ts
@@ -1,0 +1,5 @@
+declare module "*.svg" {
+  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
+  const src: string;
+  export default src;
+}
