+++ src/react-app-env.d.ts
@@ -1,0 +1,9 @@
+/// <reference types="react-scripts" />
+
+declare namespace NodeJS {
+  interface ProcessEnv {
+    NODE_ENV: "development" | "production" | "test";
+    PUBLIC_URL: string;
+    REACT_APP_TEST: string;
+  }
+}
